admin_tool/requirements.txt
customtkinter==5.2.2
supabase==2.3.4
python-dotenv==1.0.1
